clean models_parts and separate train functions to a new script file
