+++ models_parts.py
@@ -4,18 +4,7 @@ import torch.nn.functional as F
 import time
 from collections import defaultdict
 
-def double_conv(in_channels, out_channels):
-    return nn.Sequential(
-        conv_relu(in_channels, out_channels, 3, 1),
-        conv_relu(out_channels, out_channels, 3, 1)
-    )
 
-def conv_relu(in_channels, out_channels, kernel, padding):
-  return nn.Sequential(
-    nn.Conv2d(in_channels, out_channels, kernel, padding=padding),
-    nn.ReLU(inplace=True)
-  )
-  
 class DoubleConv(nn.Module):
     """(convolution => [Batch Normalization] => ReLU) * 2"""
 
@@ -55,7 +44,7 @@ class Up(nn.Module):
     def __init__(self, in_channels, out_channels, bilinear=True):
         super().__init__()
 
-        # if bilinear, use the normal convolutions to reduce the number of channels
+        # si bilinéaire, utilisez les convolutions normales pour réduire le nombre de channels
         if bilinear:
             self.up = nn.Upsample(scale_factor=2, mode='bilinear', align_corners=True)
             self.conv = DoubleConv(in_channels, out_channels, in_channels // 2)
@@ -66,7 +55,7 @@ class Up(nn.Module):
 
     def forward(self, x1, x2):
         x1 = self.up(x1)
-        # input is C H W
+        # input est C H W
         diffY = x2.size()[2] - x1.size()[2]
         diffX = x2.size()[3] - x1.size()[3]
 
@@ -83,100 +72,8 @@ class OutConv(nn.Module):
     def forward(self, x):
         return self.conv(x)
 
-# Source: A survey of loss functions for semantic segmentation https://arxiv.org/pdf/2006.14822.pdf
-def dice_loss(pred, target, smooth = 1.):
-    pred = pred.contiguous()
-    target = target.contiguous()    
-
-    intersection = (pred * target).sum(dim=2).sum(dim=2)
-    
-    loss = (1 - ((2. * intersection + smooth) / (pred.sum(dim=2).sum(dim=2) + target.sum(dim=2).sum(dim=2) + smooth)))
-    
-    return loss.mean()
-
-
-def calc_loss(pred, target, metrics, bce_weight=0.5):
-    bce = F.binary_cross_entropy_with_logits(pred, target)
-
-    pred = torch.sigmoid(pred)
-    dice = dice_loss(pred, target)
-
-    loss = bce * bce_weight + dice * (1 - bce_weight)
-
-    metrics['bce'] += bce.data.cpu().numpy() * target.size(0)
-    metrics['dice'] += dice.data.cpu().numpy() * target.size(0)
-    metrics['loss'] += loss.data.cpu().numpy() * target.size(0)
-
-    return loss
-
-def print_metrics(metrics, epoch_samples, phase):
-    outputs = []
-    for k in metrics.keys():
-        outputs.append("{}: {:4f}".format(k, metrics[k] / epoch_samples))
-
-    print("{}: {}".format(phase, ", ".join(outputs)))
-
-
-def train_model(model, optimizer, scheduler, num_epochs = 25, checkpoint_path = "checkpoint.pt"):
-    best_loss = 1e10
-
-    for epoch in range(num_epochs):
-        print('Epoch {}/{}'.format(epoch, num_epochs - 1))
-        print('-' * 10)
-
-        since = time.time()
-
-        # Each epoch has a training and validation phase
-        for phase in ['train', 'val']:
-            if phase == 'train':
-                model.train()  # Set model to training mode
-            else:
-                model.eval()   # Set model to evaluate mode
-
-            metrics = defaultdict(float)
-            epoch_samples = 0
-
-            for inputs, (labels, bounding_box) in dataloaders[phase]:
-                
-                inputs = inputs.to(device)
-                labels = labels.to(device)
-
-                # zero the parameter gradients
-                optimizer.zero_grad()
-
-                # forward
-                # track history if only in train
-                with torch.set_grad_enabled(phase == 'train'):
-                    outputs = model(inputs)
-                    loss = calc_loss(outputs, labels, metrics)
-
-                    # backward + optimize only if in training phase
-                    if phase == 'train':
-                        loss.backward()
-                        optimizer.step()
-
-                # statistics
-                epoch_samples += inputs.size(0)
-
-            print_metrics(metrics, epoch_samples, phase)
-            epoch_loss = metrics['loss'] / epoch_samples
-
-            if phase == 'train':
-              scheduler.step()
-              for param_group in optimizer.param_groups:
-                  print("LR", param_group['lr'])
-
-            # save the model weights
-            if phase == 'val' and epoch_loss < best_loss:
-                print(f"saving best model to {checkpoint_path}")
-                best_loss = epoch_loss
-                torch.save(model.state_dict(), checkpoint_path)
-
-        time_elapsed = time.time() - since
-        print('{:.0f}m {:.0f}s'.format(time_elapsed // 60, time_elapsed % 60))
-
-    print('Best val loss: {:4f}'.format(best_loss))
-
-    # load best model weights
-    model.load_state_dict(torch.load(checkpoint_path))
-    return model
+def conv_relu(in_channels, out_channels, kernel, padding):
+  return nn.Sequential(
+    nn.Conv2d(in_channels, out_channels, kernel, padding=padding),
+    nn.ReLU(inplace=True)
+  )
